spirv-opt: Run tests second time with lower max id bound
The SinglePassRunAndCheck function duplicated logic from
SinglePassRunAndDisassemble. This change refactors the former to call the
latter, simplifying the implementation and improving maintainability.

This allows us to have a central place where we can run tests a
second time with a lower id bound that will guarentee an id overflow. We
can test that the pass does not crash.

This does not necessarily cover all code paths, but should cover most.